requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.25.1
environ==1.0
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
phonenumbers==8.13.6
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
